Move empirical_influence
