+++ src/resample/bootstrap.py
@@ -19,28 +19,6 @@ from resample.jackknife import jackknife
 from resample.utils import eqf
 
 
-def empirical_influence(a: np.ndarray, f: Callable) -> np.ndarray:
-    """
-    Calculate the empirical influence function for a given sample and estimator
-    using the jackknife method.
-
-    Parameters
-    ----------
-    a : array-like
-        Sample. Must be one-dimensional.
-
-    f : callable
-        Estimator. Can be any mapping ℝⁿ → ℝᵏ, where n is the sample size
-        and k is the length of the output array.
-
-    Returns
-    -------
-    np.ndarray
-        Empirical influence values.
-    """
-    return (len(a) - 1) * (f(a) - jackknife(a, f))
-
-
 def bootstrap(
     a: np.ndarray,
     f: Optional[Callable] = None,
